extracted getCardsWithInfo controller
